system/perception/requirements.txt
opencv-python
numpy
pyserial
pyyaml
pytest
